FilePatcher.PatchAll to not iterate over all the files.
- The very beginning of the method already does all of the work necessary when it finds all fully qualified '.bpf' file paths.
- Removed recursive directory crawling to instead loop through the succinct list of known files already found.
- The trick behind this commit is utilizing 'file.FullName.Substring(patchpath.Length)...' in order to get at the 'relative' part of the patch file path.
  ^ Then it's just a matter of duct taping targetpath and the new relative path together in order to get the official target FileInfo object.

NOTE: I didn't understand the di.Refresh() followed by the removal of empty folders. That code was removed mainly because the method is no longer recursive and thus would only delete a truly empty patch directory, which seems harmless to me. There is a chance that this commit caused a loss of functionality with directory removal that I did not fully understand.